requirements.txt
aiofiles==22.1.0
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.4.0
browser-cookie3==0.19.1
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
crack==0.7
f2==0.0.1.6
fonttools==4.53.0
gmssl==3.2.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib-resources==6.1.0
iniconfig==2.0.0
iso8601==2.1.0
jsonpath-ng==1.6.0
lz4==4.3.3
m3u8==3.6.0
markdown-it-py==3.0.0
mdurl==0.1.2
packaging==24.1
pip-system-certs==4.0
pluggy==1.5.0
ply==3.11
protobuf==4.23.0
pycryptodome==3.20.0
pycryptodomex==3.20.0
pydantic==2.6.4
pydantic_core==2.16.3
PyExecJS==1.5.1
Pygments==2.18.0
pypng==0.20220715.0
pysodium==0.7.17
pytest==8.2.1
pytest-asyncio==0.21.1
PyYAML==6.0.1
qrcode==7.4.2
rel==0.4.9.19
requests==2.31.0
rich==13.7.1
six==1.16.0
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==2.2.1
websocket-client==1.6.2
websockets==12.0
wrapt==1.16.0
